+++ Releases/Utility Mod/Changelog.md
@@ -2,6 +2,7 @@
 - Added option to go to custom campaigns from right click menu.
 - Added option to go to custom campaign on startup.
 - Prevent vehicles from spawning when enemy spawns are disabled.
+- Fix always show mouse cursor option causing mouse to flicker instead.
 
 ## v2.4.0
 - Added option to set keybindings for pretty much all toggles and buttons.